sakhi_backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
symspellpy
language-tool-python
langdetect
langchain
openai
langchain_openai